f64 -> f32

+++ src/camera.rs
@@ -27,7 +27,7 @@ impl Camera {
     }
 
     #[inline(always)]
-    pub fn get_ray(&self, u: f64, v: f64) -> Ray {
+    pub fn get_ray(&self, u: f32, v: f32) -> Ray {
         Ray::new(
             self.origin.clone(),
             &self.lower_left_corner + &(u * &self.horizontal) + v * &self.vertical
