+++ src/config.rs
@@ -2,26 +2,26 @@ use std::sync::Arc;
 
 use once_cell::sync::Lazy;
 
+use crate::{color, point3, x, y, z};
 use crate::material::{Lambertian, Metal};
 use crate::prelude::*;
-use crate::{color, point3, x, y, z};
 
 // IMAGE
-pub static ASPECT_RATIO: f64 = 16.0 / 9.0;
+pub static ASPECT_RATIO: f32 = 16.0 / 9.0;
 // pub static HI_RES: bool = cfg!(debug_assertions); // for testing with AMD uPerf
 pub static HI_RES: bool = cfg!(not(debug_assertions));
 // normal operation
 pub static IMAGE_WIDTH: u64 = if HI_RES { 1920 } else { 256 };
-pub static IMAGE_HEIGHT: u64 = (IMAGE_WIDTH as f64 / ASPECT_RATIO) as u64;
+pub static IMAGE_HEIGHT: u64 = (IMAGE_WIDTH as f32 / ASPECT_RATIO) as u64;
 pub static SAMPLES_PER_PIXEL: u64 = 2 << 6;
 pub static MAX_DEPTH: u64 = 5;
 pub static OUTPUT_FILE: &str = if HI_RES { "hi_res.ppm" } else { "img.ppm" };
 pub static RENDER_FILE: &str = if HI_RES { "hi_res.png" } else { "img.png" };
 
 // CAMERA
-pub static VIEWPORT_HEIGHT: f64 = 2.0;
-pub static VIEWPORT_WIDTH: f64 = VIEWPORT_HEIGHT * ASPECT_RATIO;
-pub static FOCAL_LENGTH: f64 = 1.0;
+pub static VIEWPORT_HEIGHT: f32 = 2.0;
+pub static VIEWPORT_WIDTH: f32 = VIEWPORT_HEIGHT * ASPECT_RATIO;
+pub static FOCAL_LENGTH: f32 = 1.0;
 pub static ORIGIN: &Point3 = &point3![0];
 pub static HORIZONTAL: &Point3 = &x!(VIEWPORT_WIDTH);
 pub static VERTICAL: &Point3 = &y!(VIEWPORT_HEIGHT);
@@ -33,7 +33,7 @@ pub static LOWER_LEFT_CORNER: &Point3 = &point3![
 // &(ORIGIN - HORIZONTAL / 2.0 - VERTICAL / 2.0 + z!(FOCAL_LENGTH));
 
 // CONSTANTS
-pub static NEAR_ZERO_EPSILON: f64 = 1.0e-8;
+pub static NEAR_ZERO_EPSILON: f32 = 1.0e-8;
 
 // MATERIALS
 pub static GROUND_MATERIAL: Lazy<Arc<Lambertian>> =
