+++ src/point3_macros.rs
@@ -2,12 +2,12 @@
 macro_rules! point3 {
     ($l: expr) => {{
         Point3 {
-            0: [$l as f64, $l as f64, $l as f64],
+            0: [$l as f32, $l as f32, $l as f32],
         }
     }};
     ($l0: expr, $l1: expr, $l2: expr) => {{
         Point3 {
-            0: [$l0 as f64, $l1 as f64, $l2 as f64],
+            0: [$l0 as f32, $l1 as f32, $l2 as f32],
         }
     }};
 }
@@ -15,7 +15,7 @@ macro_rules! point3 {
 #[macro_export]
 macro_rules! x {
     ($x: expr) => {{
-        $crate::point3![1_f64 * $x as f64, 0, 0]
+        $crate::point3![1_f32 * $x as f32, 0, 0]
     }};
     () => {{
         x!(1)
@@ -25,7 +25,7 @@ macro_rules! x {
 #[macro_export]
 macro_rules! y {
     ($y: expr) => {{
-        $crate::point3![0, 1_f64 * $y as f64, 0]
+        $crate::point3![0, 1_f32 * $y as f32, 0]
     }};
     () => {{
         y!(1)
@@ -35,7 +35,7 @@ macro_rules! y {
 #[macro_export]
 macro_rules! z {
     ($z: expr) => {{
-        $crate::point3![0, 0, -1_f64 * $z as f64]
+        $crate::point3![0, 0, -1_f32 * $z as f32]
     }};
     () => {{
         z!(1)
