+++ src/sphere.rs
@@ -8,7 +8,7 @@ use crate::value;
 #[derive(Clone, Debug)]
 pub struct Sphere {
     center: Point3,
-    radius: f64,
+    radius: f32,
     material: Arc<dyn Scatter>,
 }
 
@@ -18,7 +18,7 @@ unsafe impl Sync for Sphere {}
 
 impl Sphere {
     #[inline(always)]
-    pub fn new(center: Point3, radius: f64, material: Arc<dyn Scatter>) -> Self {
+    pub fn new(center: Point3, radius: f32, material: Arc<dyn Scatter>) -> Self {
         Sphere {
             center,
             radius,
@@ -29,7 +29,7 @@ impl Sphere {
 
 impl Hittable for Sphere {
     #[inline(always)]
-    fn hit(&self, ray: &Ray, t_min: f64, t_max: f64) -> Option<HitRecord> {
+    fn hit(&self, ray: &Ray, t_min: f32, t_max: f32) -> Option<HitRecord> {
         let oc = &ray.origin - &self.center;
         let r_direction = ray.get_direction_denormalized();
